Change Cassandra to use quorum consistency level
This change was motivated by a strange situation: a network partition in a multi-DC cluster, where 2 specific DCs
couldn't talk to each other. This is a workaround while we resolve, enabling us to at least deploy. Schema changes
should still work applied anywhere, but may report a failure if initially run in one of those 2 DCs (I tried this out).